requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
psutil>=5.8.0
joblib>=1.1.0
xgboost>=1.5.0
catboost>=1.0.0
pyyaml>=5.4.0
jupyter>=1.0.0
notebook>=6.4.0